requirements.txt
asgiref
dj-database-url
Django
django-modeltranslation
pillow
psycopg2-binary
sqlparse
typing_extensions
gunicorn
bs4
python-dotenv
easy-thumbnails
